Revert "ドツボに嵌っている"
This reverts commit 198a7cf467682779cd6c932270d7642cd4210381.

+++ Project03/GameBoard.cpp
@@ -477,7 +477,7 @@ VECTOR2 GameBoard::SetListSerch(PIECE_ST type)
 		return { -1,-1 };
 	}
 
-	int rnd = GetRand(static_cast<int>(setlist.size()));
+	int rnd = GetRand(setlist.size());
 	int cnt = 0;
 	for (auto itr : setlist[static_cast<int>(type)])
 	{
